Port pspgl to PSPSDK.
Remove the sce* headers and stubs as part of the port to PSPSDK.

Move module-specific initialization code out of pspgl and into the samples.
 
Fixup floating point constants in the samples, and add
-fsingle-precision-constant to the pspgl Makefile to remove doubles.

Makefile: Add -G0 to CFLAGS to disable the $gp register.  Put autogenerated
dependencies in a seperate subdirectory to reduce clutter.  Create GL if it
doesn't exist (need to find a way to do this on platforms that don't support
symbolic links).