This implements proper object lifetime management.
In general, any in-memory object which can be pointed to by the
command queue/hardware must be reference counted.  When a command is
inserted into the queue which refers to the object, its refcount must
be incremented; once the hardware has finished with it (ie, the last
command referring to the object has been completed), the reference can
be removed.

This patch implements such a mechanism; each dlist buffer has a list
of cleanup functions which are to be run once that command buffer has
been completed.  Currently the lifetimes of texture images and colour
maps are handled properly; in future vertex buffers, framebuffer and
so on will be handled in the same way.  The callback could be used in
a more general way to simply get a notification callback once a
particular command has been executed.