This patch implements fixed-size matrix stacks.  The variable sized stacks are not useful in OpenGL, and make matrix stack operations very expensive.
In addition, this patch maintains current matrix and current matrix
stack pointers, which makes access to the current matrix state very
efficient.

It also fixes bugs in the previous code.  For some reason, all code
was using [depth-1] to index the matrix stack, but depth was
initialized to 0, causing bad memory accesses.

This change does not end up changing the generated code size.