This fixes up lighting in the following ways:
1. OpenGL has a couple of implicit ways of representing state which
need to be explicitly set in the PSP hardware.  If you set a light
with w=0 homogeneous coordinate (ie x,y,z,0), then the light becomes a
directional light, and x,y,z are a direction vector; otherwise it is a
positional light.  Also, if you set SPOT_CUTOFF to 180deg, the light
becomes a point light, otherwise its a spotlight.  Directional, point
and spot lights are exclusive, and the GE has a single "light type"
register for setting this.

2. The light type register also has a "kind" field which is some
function of which light components are currently in use.  I compute
this by looking at the current set of non-zero colours for a particular
light.

3. The current colour (set with glColor()) and the material ambient
colour (set with glMaterial(, GL_AMBIENT, )) share the same hardware
register.  The current colour is only used with lighting disabled, and
the material ambient is only use when lighting is enabled.  I
therefore switch the appropriate colour in when
glEnable/Disable(GL_LIGHTING) is called, and only update the hardware
in glColor/glMaterial if the right GL_LIGHTING enable state is
current.

4. OpenGL defines default initial material colours, so make sure
they're initialized properly.