Make pspgum_vfpu use the pspvfpu library for VFPU state management. The gum* functions use it without allocating a vfpu context, which means they won't interfere with other VFPU users but won't incur much overhead. However, the calls to pspvfpu_use_matrices() will be a little more efficient if a context is allocated, because it can use the context to keep track of the temp matrices it uses, rather than having to scan the context array each time.
The sceGum* functions only allocate a VFPU context on sceGumLoadMatrix
and sceGumLoadIdentity; that's because until either of those functions
are called, the state of the respective matrix stack is undefined, and
any operations on them are basically garbage.  If the VFPU context is
NULL, then the sceGum*() functions will produce junk, but at least they
won't interfere with any other VFPU users.

I also added the gumInit() function.  This is basically optional,
and can be called more than once.  For non-VFPU GUM it does nothing.
For VFPU GUM it allocates a vfpu context, which has the side effect of
setting the VFPU thread attribute, which is necessary for any VFPU use
(a caller could set it for itself rather than calling gumInit()).